requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.2
plotly>=5.20.0
kaleido>=0.2.1
weasyprint>=61.0
Jinja2>=3.1.3
psycopg2-binary>=2.9.9
